register pytest options in a rootdir conftest so bare pytest invocations see them
