+++ yt_idv/tests/conftest.py
@@ -1,78 +1,16 @@
 import base64
-import os
 
 import pytest
 import yt
 from pytest_html import extras as html_extras
 
 import yt_idv
-from yt_idv.rendering_contexts import prepare_platform
 
-BACKENDS = ("osmesa", "egl", "pyglet")
+# the CLI options and pytest_configure/pytest_report_header hooks that
+# select the backend and canvas size live in the rootdir conftest.py
 
 # a hidden pyglet window still needs to be told not to show itself or build a gui
 _BACKEND_KWARGS = {"pyglet": {"visible": False, "gui": False}}
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--offscreen-backend",
-        choices=BACKENDS,
-        default=None,
-        help="offscreen rendering backend to run the test suite against",
-    )
-    parser.addini(
-        "offscreen_backend",
-        help="default offscreen rendering backend, if --offscreen-backend is unset",
-        default="osmesa",
-    )
-    parser.addoption(
-        "--canvas-width",
-        type=int,
-        default=1024,
-        help="default width of offscreen rendering canvases",
-    )
-    parser.addoption(
-        "--canvas-height",
-        type=int,
-        default=1024,
-        help="default height of offscreen rendering canvases",
-    )
-
-
-def _resolve_backend(config):
-    backend = config.getoption("--offscreen-backend")
-    if backend is not None:
-        return backend
-
-    source = "YT_IDV_TEST_BACKEND"
-    backend = os.environ.get(source)
-    if backend is None:
-        source = "offscreen_backend"
-        backend = config.getini(source) or "osmesa"
-
-    if backend not in BACKENDS:
-        raise pytest.UsageError(f"{source}={backend!r} is not one of {BACKENDS}")
-    return backend
-
-
-def pytest_configure(config):
-    # this will get run before all tests, before collection and
-    # any opengl imports that happen within test files.
-    config._offscreen_backend = _resolve_backend(config)
-    config._canvas_size = (
-        config.getoption("--canvas-width"),
-        config.getoption("--canvas-height"),
-    )
-    prepare_platform(config._offscreen_backend)
-
-
-def pytest_report_header(config):
-    width, height = config._canvas_size
-    return (
-        f"offscreen backend: {config._offscreen_backend}, "
-        f"default canvas size: {width}x{height}"
-    )
 
 
 @pytest.fixture(scope="session")
